requirements.txt
alembic==1.13.1
aniso8601==9.0.1
astroid==3.1.0
attrs==23.2.0
certifi==2024.2.2
chardet==5.2.0
click==8.1.7
ConfigArgParse==1.7
coverage==7.4.3
dominate==2.9.1
Flask==3.0.2
Flask-BasicAuth==0.2.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-paginate==2024.4.12
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
lazy-object-proxy==1.10.0
locust==2.24.0
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
msgpack==1.0.7
psutil==5.9.8
PyJWT==2.8.0
pylint==3.1.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.8.2
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.27
typed-ast==1.5.5
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
